Refactor board.model->board.domain, posting.domain 하위 패키지로 구분
