+++ board/src/main/java/com/tubeplus/board_service/adapter/web/controller/board/vo/ReqMakeBoardBody.java
@@ -1,6 +1,6 @@
 package com.tubeplus.board_service.adapter.web.controller.board.vo;
 
-import com.tubeplus.board_service.board.model.BoardType;
+import com.tubeplus.board_service.board.domain.BoardType;
 import com.tubeplus.board_service.board.port.in.BoardUseCase;
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.Min;
