+++ board/src/main/java/com/tubeplus/board_service/adapter/web/controller/board/vo/VoBoardProperty.java
@@ -1,7 +1,7 @@
 package com.tubeplus.board_service.adapter.web.controller.board.vo;
 
-import com.tubeplus.board_service.board.model.Board;
-import com.tubeplus.board_service.board.model.BoardType;
+import com.tubeplus.board_service.board.domain.Board;
+import com.tubeplus.board_service.board.domain.BoardType;
 import com.tubeplus.board_service.board.port.in.BoardUseCase;
 import lombok.Builder;
 import lombok.Data;
