+++ board/src/main/java/com/tubeplus/board_service/adapter/web/controller/posting/PostingController.java
@@ -5,8 +5,8 @@ import com.tubeplus.board_service.adapter.web.common.ApiTag;
 import com.tubeplus.board_service.adapter.web.controller.posting.vo.posting.*;
 import com.tubeplus.board_service.adapter.web.error.BusinessException;
 import com.tubeplus.board_service.adapter.web.error.ErrorCode;
-import com.tubeplus.board_service.posting.domain.Posting;
-import com.tubeplus.board_service.posting.domain.PostingViewInfo;
+import com.tubeplus.board_service.posting.domain.posting.Posting;
+import com.tubeplus.board_service.posting.domain.posting.PostingViewInfo;
 import com.tubeplus.board_service.posting.port.in.PostingUseCase;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
