+++ board/src/main/java/com/tubeplus/board_service/adapter/web/controller/posting/vo/vote/VoPostingVote.java
@@ -1,8 +1,8 @@
 package com.tubeplus.board_service.adapter.web.controller.posting.vo.vote;
 
 
-import com.tubeplus.board_service.posting.domain.PostingVote;
-import com.tubeplus.board_service.posting.domain.PostingVoteType;
+import com.tubeplus.board_service.posting.domain.vote.PostingVote;
+import com.tubeplus.board_service.posting.domain.vote.PostingVoteType;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
