+++ board/src/main/java/com/tubeplus/board_service/board/domain/Board.java
@@ -1,4 +1,4 @@
-package com.tubeplus.board_service.board.model;
+package com.tubeplus.board_service.board.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
